refactoring: restructuring of the code into modules
