+++ src/function_code_handling.py
@@ -1,0 +1,84 @@
+"""This module contains functions that handle functions' decompiled code"""
+from collections import OrderedDict
+
+STACK_PROTECTOR_VARIABLE = "in_FS_OFFSET"
+TYPES_TO_REPLACE = OrderedDict(uint="unsigned int",
+                               ushort="unsigned short",
+                               ulong="unsigned long",
+                               undefined2="uint16_t",
+                               undefined3="uint32_t",
+                               undefined5="uint64_t",
+                               undefined6="uint64_t",
+                               undefined7="uint64_t")
+
+
+def replace_types(code):
+    """Replacing all Ghidra types with types from intttypes.h and standart C types"""
+    for old_type, new_type in TYPES_TO_REPLACE.items():
+        code = code.replace(old_type, new_type)
+    return code
+
+
+def remove_stack_protection(code):
+    """Removes stack protection from function code"""
+    lines = code.split('\n')
+    for num, line in enumerate(lines):
+        if STACK_PROTECTOR_VARIABLE in line:
+            # if we have "if ..." with in_FS_OFFSET checking
+            # we must remove all "if" block - 4 lines (with ghidra comment)
+            if "if" in line:
+                # if == is in the condition:
+                # lines[num] = "if ... == ... STACK_PROTECTOR_VARIABLE {"
+                # lines[num + 1] = "  return 0;"
+                # lines[num + 2] = "}"
+                # lines[num + 3] is Ghidra's comment
+                # lines[num + 4] = "__stack_chk_fail();"
+                # if != is in the condition:
+                # lines[num] = "if ... != ... STACK_PROTECTOR_VARIABLE {"
+                # lines[num + 1] is Ghidra's comment
+                # lines[num + 2] = "__stack_chk_fail();"
+                # lines[num + 3] = "}"
+                # lines[num + 4] = "return 0;" - we keep it
+                if "==" in line:
+                    lines[num + 1] = lines[num + 1][2:]
+                    lines.pop(num + 4)
+                lines.pop(num + 3)
+                lines.pop(num + 2)
+                if "!=" in line:
+                    lines.pop(num + 1)
+            lines.pop(num)
+    new_code = '\n'.join(lines)
+    return new_code
+
+
+def line_from_body(line, signature):
+    """Line is from function body if it is not a comment, is not empty, 
+    is not a { or } and is not its signature. Function checks that line is from body"""
+    return not (line.startswith(("//", "/*")) or line == ''
+                or line in "{}" or line == signature[:-1])
+
+
+def is_single_return(code, signature):
+    """If function body consists of single return;, it is service function. 
+    Function checks if function consists of single return"""
+    body = [line.replace(' ', '') for line in code.split('\n') if line_from_body(line, signature)]
+    return len(body) == 1 and body[0] == "return;"
+
+
+def calls_single_return(code, signature, single_return_functions):
+    """If function calls single return function, it is service function. 
+    Function checks if function calls single return function."""
+    body = [line.replace(' ', '') for line in code.split('\n') if line_from_body(line, signature)]
+    for function in single_return_functions:
+        if function + "();" in body:
+            return True
+    return False
+
+
+def handle_function(code):
+    """Handling function code"""
+    code_replaced_types = replace_types(code)
+    if STACK_PROTECTOR_VARIABLE not in code_replaced_types:
+        return code_replaced_types
+    code_removed_stack_protection = remove_stack_protection(code_replaced_types)
+    return code_removed_stack_protection
