+++ src/function_handling.py
@@ -1,0 +1,33 @@
+"""This module contains functions that handle functions before decompiling"""
+
+HEX_BASE = 16
+RUNTIME_PREFIX = '_'
+PLT_INSTRUCTION = "JMP qword ptr"
+
+
+def address_to_int(address):
+    """Converts address (hex number) to int"""
+    return int(str(address), HEX_BASE)
+
+
+def function_in_runtime(function):
+    """Checks if input function is from C Runtime"""
+    function_name = function.getName()
+    return function_name.startswith(RUNTIME_PREFIX)
+
+
+def function_is_plt(function):
+    """Checks if input function is PLT jump"""
+    program = function.getProgram()
+    listing = program.getListing()
+    body = function.getBody()
+    min_address = address_to_int(body.getMinAddress())
+    max_address = address_to_int(body.getMaxAddress())
+    for address in body.getAddresses(True):
+        code_unit = str(listing.getCodeUnitAt(address))
+        if code_unit.startswith(PLT_INSTRUCTION):
+            words = code_unit.split()
+            jmp_address = address_to_int(words[-1][1:-1])  # [1:-1] is to remove [] from address
+            if not min_address <= jmp_address <= max_address:
+                return True
+    return False
