+++ src/main.py
@@ -2,8 +2,9 @@
 
 # pylint: disable=wrong-import-position, import-error, wrong-import-order
 from shutil import rmtree
-import pyhidra
 import tools
+import global_variables_handling
+import pyhidra
 from java.io import File, PrintWriter
 
 LIBRARY_LIST = ("stdio.h", "stdlib.h", "inttypes.h", "stdbool.h")
@@ -21,7 +22,7 @@ def export_c_code(binary_file_path, output_file_path):
             c_file_writer.println(f"#include <{lib}>")
         tools.write_program_data_types(program, c_file_writer, flat_api.monitor)
         for section in SECTIONS:
-            tools.write_global_variables(program, c_file_writer, section)
+            global_variables_handling.write_global_variables(program, c_file_writer, section)
             c_file_writer.println()
         tools.put_functions(program, c_file_writer, flat_api.monitor)
         c_file_writer.close()
@@ -29,4 +30,4 @@ def export_c_code(binary_file_path, output_file_path):
     rmtree(f"resources/in/{project_folder}")
 
 
-export_c_code("resources/in/avl", "resources/out/avl5.c")
+export_c_code("resources/in/hello_world.out", "resources/out/test.c")
